Change reset button color and french label
Given that the 'reset' button has a blue color, we can think that it is a
validate button.  Since it has a destructive effect (clear inputs), I think
that another color could be a better choice.

I propose to use `bnt-default' as new color.
I've also updated the french label, so it effect is more explicit in my
opinion.

This commit is licencied under the MIT Licence, to be easily included in
bnjbvr/kresus repository.